chapter_title/requirements.txt
pika
psycopg2
pdfplumber
Levenshtein
python-dotenv
google-cloud-storage